requirements.txt
torch
diffusers["torch"]
transformers
numpy==1.20.0
Pillow